requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-CORS
APScheduler
beautifulsoup4
requests
python-dotenv